perf: apply advanced Vercel React best practices
Apply the three future recommendations from the review:

1. React.memo (rerender-memo)
   - Wrap OrderTable with memo() to prevent unnecessary re-renders
   - Extract OrderRow as memoized component for row-level optimization

2. content-visibility CSS (rendering-content-visibility)
   - Add content-visibility: auto to table rows for virtualization-like perf
   - Add containIntrinsicSize for proper layout calculation
   - Apply to order items list in detail page

3. Suspense boundaries (async-suspense-boundaries)
   - Split order detail page into main content and sidebar components
   - Add Suspense wrapper for sidebar that fetches delivery persons
   - Create skeleton fallbacks for loading states
   - Enable streaming of independent page sections
